+++ rich-html-editor/src/main/java/com/infomaniak/lib/richhtmleditor/executor/SpellCheckHtmlSetter.kt
@@ -19,15 +19,15 @@ package com.infomaniak.lib.richhtmleditor.executor
 
 import android.webkit.WebView
 import com.infomaniak.lib.richhtmleditor.RichHtmlEditorWebView.Companion.EDITOR_ID
+import com.infomaniak.lib.richhtmleditor.encodeArgsForJs
 
 internal class SpellCheckHtmlSetter(private val webView: WebView) : JsLifecycleAwareExecutor<Boolean>() {
 
     override fun executeImmediately(value: Boolean) = webView.insertSpellCheckValue(value)
 
     private fun WebView.insertSpellCheckValue(enable: Boolean) {
-        val spellCheckValue = if (enable) "true" else "false"
         evaluateJavascript(
-            """document.getElementById("$EDITOR_ID").setAttribute("spellcheck", $spellCheckValue)""",
+            """document.getElementById("$EDITOR_ID").setAttribute("spellcheck", ${encodeArgsForJs(enable)})""",
             null
         )
     }
